Refactor "add assets" flow to go through intro screen
