+++ src/navigator/Screens.tsx
@@ -37,6 +37,7 @@ export enum Screens {
   ImportWallet = 'ImportWallet',
   Invite = 'Invite',
   JumpstartEnterAmount = 'JumpstartEnterAmount',
+  JumpstartIntroScreen = 'JumpstartIntroScreen',
   JumpstartSendConfirmation = 'JumpstartSendConfirmation',
   JumpstartShareLink = 'JumpstartShareLink',
   KeylessBackupPhoneInput = 'KeylessBackupPhoneInput',
