+++ test/RootStateSchema.json
@@ -3313,6 +3313,7 @@
                 "ImportWallet",
                 "Invite",
                 "JumpstartEnterAmount",
+                "JumpstartIntroScreen",
                 "JumpstartSendConfirmation",
                 "JumpstartShareLink",
                 "JumpstartTransactionDetailsScreen",
